backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pymysql==1.1.1
alembic==1.13.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pymilvus==2.4.4
python-multipart==0.0.9
pydantic-settings==2.5.2
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
anyio==4.6.2
pdfplumber==0.11.4
python-docx==1.1.2
python-pptx==1.0.2
FlagEmbedding==1.2.11
celery==5.4.0
redis==5.1.1
openpyxl==3.1.5
feedparser==6.0.11
beautifulsoup4==4.12.3
crawl4ai[all]
apscheduler==3.10.4
lark-oapi==1.4.0
pycryptodome==3.21.0
matplotlib>=3.8
funasr
Pillow
openai

langchain-openai>=0.3.0
langchain-text-splitters>=0.3.0
langchain-core>=0.3.0
whisperlivekit
